Minimise bigint.h
